move a method
