+++ step_setup.go
@@ -4,6 +4,10 @@ import (
 	"github.com/mitchellh/multistep"
 	"github.com/hashicorp/packer/packer"
 	"github.com/vmware/govmomi/find"
+	"fmt"
+	"github.com/vmware/govmomi"
+	"context"
+	"net/url"
 )
 
 type StepSetup struct{
@@ -46,3 +50,24 @@ func (s *StepSetup) Run(state multistep.StateBag) multistep.StepAction {
 }
 
 func (s *StepSetup) Cleanup(state multistep.StateBag) {}
+
+func createClient(URL, username, password string) (*govmomi.Client, context.Context, error) {
+	// create context
+	ctx := context.TODO() // an empty, default context (for those, who is unsure)
+
+	// create a client
+	// (connected to the specified URL,
+	// logged in with the username-password)
+	u, err := url.Parse(URL) // create a URL object from string
+	if err != nil {
+		return nil, nil, err
+	}
+	u.User = url.UserPassword(username, password) // set username and password for automatical authentification
+	fmt.Println(u.String())
+	client, err := govmomi.NewClient(ctx, u,true) // creating a client (logs in with given uname&pswd)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return client, ctx, nil
+}
